# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Cargo Project",
        "title": "Cargo Project API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/cargos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all cargos in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargos"
                ],
                "summary": "List all cargos",
                "responses": {
                    "200": {
                        "description": "List of cargos",
                        "schema": {
                            "$ref": "#/definitions/cargo.ListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new cargo with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargos"
                ],
                "summary": "Create a new cargo",
                "parameters": [
                    {
                        "description": "Cargo object to be created",
                        "name": "cargo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cargo.Cargo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cargo successfully created",
                        "schema": {
                            "$ref": "#/definitions/cargo.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cargos/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a cargo by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargos"
                ],
                "summary": "Get a cargo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cargo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cargo found",
                        "schema": {
                            "$ref": "#/definitions/cargo.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a cargo by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargos"
                ],
                "summary": "Delete a cargo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cargo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cargo deleted",
                        "schema": {
                            "$ref": "#/definitions/cargo.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a cargo by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargos"
                ],
                "summary": "Update a cargo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cargo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cargo object to be updated",
                        "name": "cargo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cargo.Cargo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cargo updated",
                        "schema": {
                            "$ref": "#/definitions/cargo.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/trucks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all trucks in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trucks"
                ],
                "summary": "List all trucks",
                "responses": {
                    "201": {
                        "description": "List of trucks",
                        "schema": {
                            "$ref": "#/definitions/truck.ListResponse"
                        }
                    },
                    "404": {
                        "description": "Trucks not found",
                        "schema": {
                            "$ref": "#/definitions/truck.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new truck with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trucks"
                ],
                "summary": "Create a new truck",
                "parameters": [
                    {
                        "description": "Truck object to be created",
                        "name": "truck",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/truck.Truck"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cargo successfully created",
                        "schema": {
                            "$ref": "#/definitions/cargo.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cargo.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/trucks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a truck by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trucks"
                ],
                "summary": "Get a truck by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Truck ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Truck found",
                        "schema": {
                            "$ref": "#/definitions/truck.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/truck.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Truck not found",
                        "schema": {
                            "$ref": "#/definitions/truck.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cargo.Cargo": {
            "type": "object",
            "required": [
                "cargoNumber",
                "driver",
                "transportationInfo",
                "truckId"
            ],
            "properties": {
                "cargoNumber": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "driver": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "loadUnloadDate": {
                    "type": "string"
                },
                "paymentStatus": {
                    "type": "string"
                },
                "payoutAmount": {
                    "type": "number"
                },
                "payoutDate": {
                    "type": "string"
                },
                "payoutTerms": {
                    "type": "string"
                },
                "transportationInfo": {
                    "type": "string"
                },
                "truckId": {
                    "type": "integer"
                }
            }
        },
        "cargo.CreateResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/cargo.Cargo"
                },
                "message": {
                    "type": "string",
                    "example": "Груз успешно создан"
                }
            }
        },
        "cargo.DeleteResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Груз успешно удалён"
                }
            }
        },
        "cargo.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Невалидный формат JSON"
                }
            }
        },
        "cargo.GetResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/cargo.Cargo"
                },
                "message": {
                    "type": "string",
                    "example": "Данные о грузе"
                }
            }
        },
        "cargo.ListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cargo.Cargo"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Список всех грузов"
                }
            }
        },
        "truck.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Невалидный формат JSON"
                }
            }
        },
        "truck.GetResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/truck.Truck"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Машина"
                }
            }
        },
        "truck.ListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/truck.Truck"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Список всех машин"
                }
            }
        },
        "truck.Truck": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token for authenticated requests",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}